api/infrastructure/imagen_client.py
import os
import uuid
from pathlib import Path

from ...domain.ports import IImagenClient
from ...config import UPLOAD_DIR # Assuming UPLOAD_DIR is where generated images will temporarily reside

class ImagenClient(IImagenClient):
    def __init__(self):
        # For now, this is a placeholder. In a real scenario, this would interact with
        # a local text-to-image model (e.g., Stable Diffusion via Ollama) or a free API.
        pass

    def generate_image(self, prompt: str) -> str:
        """Generates a dummy image and saves it to a temporary location."""
        print(f"Simulating image generation for prompt: {prompt}")
        
        # Define the directory for generated sprites within the public folder
        generated_sprites_dir = Path("/app/Projeto_O_Cocriador/public/generated_sprites")
        generated_sprites_dir.mkdir(parents=True, exist_ok=True)

        # Create a dummy image file
        image_filename = f"generated_sprite_{uuid.uuid4().hex}.png"
        image_path = generated_sprites_dir / image_filename

        # Create a simple placeholder image (e.g., a black square)
        try:
            from PIL import Image
            img = Image.new('RGB', (128, 128), color = 'black')
            img.save(image_path)
            print(f"Dummy image saved to {image_path}")
        except ImportError:
            print("Pillow not installed. Cannot create dummy image. Please install Pillow (pip install Pillow).")
            # Create an empty file as a fallback
            with open(image_path, 'w') as f:
                f.write("Placeholder image content")

        # Return the relative path that the frontend expects
        return f"/generated_sprites/{image_filename}"