requirements.txt
fastapi
uvicorn
python-multipart
httpx
celery
redis
pydantic
instructor
python-dotenv
opencv-python-headless
Pillow
confluent-kafka
pytest
pytest-mock
pytest-cov
requests
supabase
llama-cpp-python